+++ src/qobuz/qobuzrequest.h
@@ -76,13 +76,13 @@ class QobuzRequest : public QobuzBaseRequest {
   void ArtistsReplyReceived(QNetworkReply *reply, const int limit_requested, const int offset_requested);
 
   void AlbumsReplyReceived(QNetworkReply *reply, const int limit_requested, const int offset_requested);
-  void AlbumsReceived(QNetworkReply *reply, const int artist_id_requested, const int limit_requested, const int offset_requested);
+  void AlbumsReceived(QNetworkReply *reply, const qint64 artist_id_requested, const int limit_requested, const int offset_requested);
 
   void SongsReplyReceived(QNetworkReply *reply, const int limit_requested, const int offset_requested);
-  void SongsReceived(QNetworkReply *reply, const int artist_id_requested, const QString &album_id_requested, const int limit_requested, const int offset_requested, const QString &album_artist_requested = QString());
+  void SongsReceived(QNetworkReply *reply, const qint64 artist_id_requested, const QString &album_id_requested, const int limit_requested, const int offset_requested, const QString &album_artist_requested = QString());
 
-  void ArtistAlbumsReplyReceived(QNetworkReply *reply, const int artist_id, const int offset_requested);
-  void AlbumSongsReplyReceived(QNetworkReply *reply, const int artist_id, const QString &album_id, const int offset_requested, const QString &album_artist);
+  void ArtistAlbumsReplyReceived(QNetworkReply *reply, const qint64 artist_id, const int offset_requested);
+  void AlbumSongsReplyReceived(QNetworkReply *reply, const qint64 artist_id, const QString &album_id, const int offset_requested, const QString &album_artist);
   void AlbumCoverReceived(QNetworkReply *reply, const QUrl &cover_url, const QString &filename);
 
  private:
@@ -90,16 +90,15 @@ class QobuzRequest : public QobuzBaseRequest {
   typedef QList<Param> ParamList;
 
   struct Request {
-    int artist_id = 0;
+    qint64 artist_id = 0;
     QString album_id = 0;
-    int song_id = 0;
+    qint64 song_id = 0;
     int offset = 0;
     int limit = 0;
     QString album_artist;
     QString album;
   };
   struct AlbumCoverRequest {
-    //int artist_id = 0;
     QUrl url;
     QString filename;
   };
@@ -126,16 +125,16 @@ class QobuzRequest : public QobuzBaseRequest {
   void FlushSongsRequests();
 
   void ArtistsFinishCheck(const int limit = 0, const int offset = 0, const int artists_received = 0);
-  void AlbumsFinishCheck(const int artist_id, const int limit = 0, const int offset = 0, const int albums_total = 0, const int albums_received = 0);
-  void SongsFinishCheck(const int artist_id, const QString &album_id, const int limit, const int offset, const int songs_total, const int songs_received, const QString &album_artist);
+  void AlbumsFinishCheck(const qint64 artist_id, const int limit = 0, const int offset = 0, const int albums_total = 0, const int albums_received = 0);
+  void SongsFinishCheck(const qint64 artist_id, const QString &album_id, const int limit, const int offset, const int songs_total, const int songs_received, const QString &album_artist);
 
-  void AddArtistAlbumsRequest(const int artist_id, const int offset = 0);
+  void AddArtistAlbumsRequest(const qint64 artist_id, const int offset = 0);
   void FlushArtistAlbumsRequests();
 
-  void AddAlbumSongsRequest(const int artist_id, const QString &album_id, const QString &album_artist, const int offset = 0);
+  void AddAlbumSongsRequest(const qint64 artist_id, const QString &album_id, const QString &album_artist, const int offset = 0);
   void FlushAlbumSongsRequests();
 
-  int ParseSong(Song &song, const QJsonObject &json_obj, int artist_id, QString album_id, QString album_artist, QString album, QUrl cover_url);
+  int ParseSong(Song &song, const QJsonObject &json_obj, qint64 artist_id, QString album_id, QString album_artist, QString album, QUrl cover_url);
 
   QString AlbumCoverFileName(const Song &song);
 
@@ -174,7 +173,7 @@ class QobuzRequest : public QobuzBaseRequest {
   QQueue<Request> album_songs_requests_queue_;
   QQueue<AlbumCoverRequest> album_cover_requests_queue_;
 
-  QList<int> artist_albums_requests_pending_;
+  QList<qint64> artist_albums_requests_pending_;
   QHash<QString, Request> album_songs_requests_pending_;
   QMultiMap<QUrl, Song*> album_covers_requests_sent_;
